requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
openai==1.51.2
python-dotenv==1.0.1